requirements.txt
aiohttp==3.8.6
Brotli==1.0.9
pure-protobuf==3.0.0a5
yarl==1.8.2
pygame==2.5.2
websockets==11.0.3
watchdog==3.0.0
dashscope==1.13.3
requests>=2.31.0
pydantic>=2.5.1
torch==2.1.0
transformers==4.35.2
uvicorn==0.24.0.post1
fastapi==0.104.1
loguru==0.7.2
edge-tts==7.0.1
pyvts==0.3.2
pyyaml>=6.0.1
python-dotenv>=1.0.0
azure-cognitiveservices-speech>=1.32.0
pydub>=0.25.1
python-vlc>=3.0.18122